VMR_MSL39_v4d add refseq accessions and (some) taxids
Update [species_isolates] tables with refseq accession numbers and (some) NCBI taxid's.

Pipeline for refseq taxid querying still fully functional. Additioanl taxids to be added in the future.

Note that when the taxid's for all segments are idential, only one taxid is provided. If the segments have divergent taxids, then a taxid per segment is presented in the format of a semi-colon delimited list, with optional [segment_name:] prefix.